Removed unused methods - improving encapsulation
Removed getPotentials() from ElectrostaticSystem and
getBoundaryConditionPositions() for UnsolvedElectrostaticSystem and
corresponding tests. These methods provided access to underlying
implementation that was not necessary.

Using the ElectrostaticSystem and UnsolvedElectrostaticSystem classes
should not require knowledge of how they work internally, but this was
needed to make use of the getPotentials() and
getBoundaryConditionPositions() methods.

Also, the way the boundary condition positions are stored is not optimal
so it may be changed at a later time. This would break any other code
that used the getBoundayPositions() method.